calendar/requirements/local.txt
alabaster==0.7.12
appdirs==1.4.3
appnope==0.1.0
argon2-cffi==19.2.0
aspy.yaml==1.3.0
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
black==19.10b0
certifi==2019.11.28
cffi==1.13.2
cfgv==2.0.1
chardet==3.0.4
Click==7.0
coreapi==2.3.3
coreschema==0.0.4
coverage==5.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.2.9
django-allauth==0.41.0
django-coverage-plugin==1.6.0
django-crispy-forms==1.8.1
django-debug-toolbar==2.1
django-environ==0.4.5
django-extensions==2.2.5
django-model-utils==4.0.0
django-redis==4.11.0
djangorestframework==3.11.0
docutils==0.15.2
entrypoints==0.3
factory-boy==2.12.0
Faker==3.0.0
flake8==3.7.9
identify==1.4.9
idna==2.8
imagesize==1.2.0
importlib-metadata==1.3.0
importlib-resources==1.0.2
ipdb==0.12.3
ipython==7.11.1
ipython-genutils==0.2.0
isort==4.3.21
itypes==1.1.0
jedi==0.15.2
Jinja2==2.10.3
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
mccabe==0.6.1
more-itertools==8.0.2
mypy==0.761
mypy-extensions==0.4.3
nodeenv==1.3.3
oauthlib==3.1.0
packaging==20.0
parso==0.5.2
pathspec==0.7.0
pexpect==4.7.0
pickleshare==0.7.5
Pillow==6.2.1
pluggy==0.13.1
pre-commit==1.20.0
prompt-toolkit==3.0.2
psycopg2-binary==2.8.4
ptyprocess==0.6.0
py==1.8.1
pycodestyle==2.5.0
pycparser==2.19
pyflakes==2.1.1
Pygments==2.5.2
pylint==2.4.4
pylint-django==2.0.13
pylint-plugin-utils==0.6
pyparsing==2.4.6
pytest==5.3.1
pytest-django==3.7.0
pytest-sugar==0.9.2
python-dateutil==2.8.1
python-slugify==4.0.0
python3-openid==3.1.0
pytz==2019.3
PyYAML==5.2
redis==3.3.11
regex==2019.12.20
requests==2.22.0
requests-oauthlib==1.3.0
six==1.13.0
snowballstemmer==2.0.0
Sphinx==2.3.0
sphinxcontrib-applehelp==1.0.1
sphinxcontrib-devhelp==1.0.1
sphinxcontrib-htmlhelp==1.0.2
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.2
sphinxcontrib-serializinghtml==1.1.3
sqlparse==0.3.0
termcolor==1.1.0
text-unidecode==1.3
toml==0.10.0
traitlets==4.3.3
typed-ast==1.4.0
typing-extensions==3.7.4.1
uritemplate==3.0.1
urllib3==1.25.7
virtualenv==16.7.9
wcwidth==0.1.8
Werkzeug==0.16.0
wrapt==1.11.2
zipp==0.6.0
